Revert "Support encoder engine"
