+++ sglang_omni/__init__.py
@@ -1,7 +1,16 @@
 # SPDX-License-Identifier: Apache-2.0
 """SGLang-Omni: Multi-stage pipeline framework for omni models."""
 
-from sglang_omni.engines.base import EchoEngine, Engine
+# Re-export from submodules for convenience
+from sglang_omni.core.types import (
+    AbortMessage,
+    CompleteMessage,
+    DataReadyMessage,
+    RequestState,
+    SHMMetadata,
+    StageInfo,
+)
+from sglang_omni.engine.base import EchoEngine, Engine
 from sglang_omni.pipeline.coordinator import Coordinator
 from sglang_omni.pipeline.input_handler import (
     AggregatedInput,
@@ -10,16 +19,6 @@ from sglang_omni.pipeline.input_handler import (
 )
 from sglang_omni.pipeline.stage import Stage
 from sglang_omni.pipeline.worker import Worker
-
-# Re-export from submodules for convenience
-from sglang_omni.proto import (
-    AbortMessage,
-    CompleteMessage,
-    DataReadyMessage,
-    RequestState,
-    SHMMetadata,
-    StageInfo,
-)
 
 __version__ = "0.1.0"
