+++ sglang_omni/core/__init__.py
@@ -1,25 +1,24 @@
 # SPDX-License-Identifier: Apache-2.0
-from .data import SHMMetadata
-from .messages import (
+from sglang_omni.core.types import (
     AbortMessage,
     CompleteMessage,
     DataReadyMessage,
+    RequestInfo,
+    RequestState,
+    SHMMetadata,
     ShutdownMessage,
+    StageInfo,
     SubmitMessage,
-    parse_message,
 )
-from .request import RequestInfo, RequestState
-from .stage import StageInfo
 
 __all__ = [
+    "RequestState",
+    "StageInfo",
+    "RequestInfo",
     "SHMMetadata",
     "DataReadyMessage",
     "AbortMessage",
     "CompleteMessage",
     "SubmitMessage",
     "ShutdownMessage",
-    "parse_message",
-    "RequestState",
-    "RequestInfo",
-    "StageInfo",
 ]
