+++ sglang_omni/core/types.py
@@ -1,10 +1,53 @@
 # SPDX-License-Identifier: Apache-2.0
-"""Control plane messages."""
+"""Shared types for SGLang-Omni pipeline."""
 
 from dataclasses import dataclass
+from enum import Enum
 from typing import Any
 
-from .data import SHMMetadata
+# === Enums ===
+
+
+class RequestState(Enum):
+    """State of a request in the pipeline."""
+
+    PENDING = "pending"
+    RUNNING = "running"
+    COMPLETED = "completed"
+    FAILED = "failed"
+    ABORTED = "aborted"
+
+
+# === Stage Info ===
+
+
+@dataclass
+class StageInfo:
+    """Information about a registered stage."""
+
+    name: str
+    control_endpoint: str  # ZMQ endpoint for receiving control messages
+
+
+# === SHM Metadata ===
+
+
+@dataclass
+class SHMMetadata:
+    """Metadata for shared memory segment."""
+
+    name: str  # SHM segment name (system-generated)
+    size: int  # Size in bytes
+
+    def to_dict(self) -> dict[str, Any]:
+        return {"name": self.name, "size": self.size}
+
+    @classmethod
+    def from_dict(cls, d: dict[str, Any]) -> "SHMMetadata":
+        return cls(name=d["name"], size=d["size"])
+
+
+# === Control Plane Messages ===
 
 
 @dataclass
@@ -42,7 +85,10 @@ class AbortMessage:
     request_id: str
 
     def to_dict(self) -> dict[str, Any]:
-        return {"type": "abort", "request_id": self.request_id}
+        return {
+            "type": "abort",
+            "request_id": self.request_id,
+        }
 
     @classmethod
     def from_dict(cls, d: dict[str, Any]) -> "AbortMessage":
@@ -85,14 +131,21 @@ class SubmitMessage:
     """Submit a new request to the entry stage."""
 
     request_id: str
-    data: Any
+    data: Any  # Initial input data
 
     def to_dict(self) -> dict[str, Any]:
-        return {"type": "submit", "request_id": self.request_id, "data": self.data}
+        return {
+            "type": "submit",
+            "request_id": self.request_id,
+            "data": self.data,
+        }
 
     @classmethod
     def from_dict(cls, d: dict[str, Any]) -> "SubmitMessage":
-        return cls(request_id=d["request_id"], data=d["data"])
+        return cls(
+            request_id=d["request_id"],
+            data=d["data"],
+        )
 
 
 @dataclass
@@ -107,8 +160,25 @@ class ShutdownMessage:
         return cls()
 
 
+# === Request Tracking ===
+
+
+@dataclass
+class RequestInfo:
+    """Tracking info for a request in the coordinator."""
+
+    request_id: str
+    state: RequestState = RequestState.PENDING
+    current_stage: str | None = None
+    result: Any = None
+    error: str | None = None
+
+
+# === Message Parsing Helper ===
+
+
 def parse_message(
-    d: dict[str, Any],
+    d: dict[str, Any]
 ) -> (
     DataReadyMessage | AbortMessage | CompleteMessage | SubmitMessage | ShutdownMessage
 ):
