+++ sglang_omni/engine/__init__.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: Apache-2.0
-from sglang_omni.engines.base import EchoEngine, Engine
+from sglang_omni.engine.base import EchoEngine, Engine
 
 __all__ = [
     "Engine",
