+++ sglang_omni/engine/base.py
