--- sglang_omni/engines/encoder/__init__.py
@@ -1,13 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-from .encoder_engine import EncoderEngine
-from .encoder_runner import EncoderRunner, Request
-from .scheduler import PendingRequest, Scheduler, SchedulerConfig
-
-__all__ = [
-    "EncoderEngine",
-    "EncoderRunner",
-    "Request",
-    "Scheduler",
-    "SchedulerConfig",
-    "PendingRequest",
-]
